+++ src/conn/tcp/sync/tls_tcp_client.rs
@@ -1,30 +1,22 @@
 use crate::protocol::util;
 use crate::ConnectParams;
 use rustls::{ClientSession, Session};
-use std::cell::RefCell;
-use std::fmt;
-use std::net::TcpStream;
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 use webpki::DNSNameRef;
 
 pub(crate) struct TlsTcpClient {
     params: ConnectParams,
-    reader: RefCell<std::io::BufReader<TlsTcpStream>>,
-    writer: RefCell<std::io::BufWriter<TlsTcpStream>>,
+    stream: Stream,
 }
-impl fmt::Debug for TlsTcpClient {
-    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+impl std::fmt::Debug for TlsTcpClient {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> Result<(), std::fmt::Error> {
         write!(f, "TlsTcpClient {{params: {:?}, ... }}", &self.params)
     }
 }
 impl TlsTcpClient {
     pub fn try_new(params: ConnectParams) -> std::io::Result<Self> {
-        let tlsstream = TlsTcpStream::try_new(&params)?;
-        Ok(Self {
-            params,
-            reader: RefCell::new(std::io::BufReader::new(tlsstream.try_clone()?)),
-            writer: RefCell::new(std::io::BufWriter::new(tlsstream)),
-        })
+        let stream = Stream::try_new(&params)?;
+        Ok(Self { params, stream })
     }
 
     pub fn connect_params(&self) -> &ConnectParams {
@@ -32,73 +24,85 @@ impl TlsTcpClient {
     }
 
     #[allow(dead_code)]
-    pub fn reconnect(&self) -> std::io::Result<()> {
-        let tlsstream = TlsTcpStream::try_new(&self.params)?;
-        self.reader
-            .replace(std::io::BufReader::new(tlsstream.try_clone()?));
-        self.writer.replace(std::io::BufWriter::new(tlsstream));
+    pub fn reconnect(&mut self) -> std::io::Result<()> {
+        self.stream = Stream::try_new(&self.params)?;
         Ok(())
     }
 
-    pub(crate) fn writer(&self) -> &RefCell<dyn std::io::Write> {
-        &self.writer
+    pub(crate) fn writer(&mut self) -> &mut dyn std::io::Write {
+        &mut self.stream
     }
 
-    pub(crate) fn reader(&self) -> &RefCell<dyn std::io::Read> {
-        &self.reader
+    pub(crate) fn reader(&mut self) -> &mut dyn std::io::Read {
+        &mut self.stream
     }
 }
 
-struct TlsTcpStream {
+struct Stream {
     is_handshaking: bool,
-    tcpstream: TcpStream,
-    tlssession: Arc<Mutex<ClientSession>>,
+    tcpstream: std::net::TcpStream,
+    client_session: ClientSession,
 }
-impl TlsTcpStream {
+impl Stream {
     pub fn try_new(params: &ConnectParams) -> std::io::Result<Self> {
-        let (tcpstream, tlssession) = connect_tcp(params)?;
+        debug!("Stream: Connecting to {:?}", params.addr());
+        let tcpstream = std::net::TcpStream::connect(params.addr())?;
+        trace!("tcpstream working");
+
+        let client_session = ClientSession::new(
+            &Arc::new(params.rustls_clientconfig()?),
+            DNSNameRef::try_from_ascii_str(params.host()).map_err(|_| {
+                util::io_error(format!("Cannot use {} for DNSNameRef", params.host()))
+            })?,
+        );
         Ok(Self {
             is_handshaking: true,
             tcpstream,
-            tlssession: Arc::new(Mutex::new(tlssession)),
-        })
-    }
-    pub fn try_clone(&self) -> std::io::Result<Self> {
-        Ok(Self {
-            is_handshaking: false,
-            tcpstream: self.tcpstream.try_clone()?,
-            tlssession: Arc::clone(&self.tlssession),
+            client_session,
         })
     }
 }
-
-fn connect_tcp(params: &ConnectParams) -> std::io::Result<(TcpStream, ClientSession)> {
-    debug!("connect_tcp(): Connecting to {:?}", params.addr());
-    let tcpstream = TcpStream::connect(params.addr())?;
-    trace!("tcpstream working");
-
-    let tlssession = ClientSession::new(
-        &Arc::new(params.rustls_clientconfig()?),
-        DNSNameRef::try_from_ascii_str(params.host())
-            .map_err(|_| util::io_error(format!("Cannot use {} for DNSNameRef", params.host())))?,
-    );
-
-    Ok((tcpstream, tlssession))
-}
-
-impl std::io::Write for TlsTcpStream {
+impl std::io::Write for Stream {
     fn write(&mut self, request: &[u8]) -> std::io::Result<usize> {
         trace!(
             "std::io::Write::write() with request size {}",
             request.len()
         );
-        let mut tlssession = self.tlssession.lock().unwrap();
 
-        let result = std::io::Write::write(&mut *tlssession, request)?;
+        let result = std::io::Write::write(&mut self.client_session, request)?;
 
-        while tlssession.wants_write() {
-            let count = tlssession.write_tls(&mut self.tcpstream)?;
+        while self.client_session.wants_write() {
+            let count = self.client_session.write_tls(&mut self.tcpstream)?;
             trace!("std::io::Write::write(): wrote tls ({})", count);
+        }
+
+        if self.is_handshaking {
+            loop {
+                while self.client_session.wants_write() {
+                    let count = self.client_session.write_tls(&mut self.tcpstream)?;
+                    trace!("std::io::Write::write(): wrote tls ({})", count);
+                }
+                if self.is_handshaking && !self.client_session.is_handshaking() {
+                    self.is_handshaking = false;
+                    if let Some(protocol) = self.client_session.get_protocol_version() {
+                        debug!("write: Protocol {:?} negotiated", protocol)
+                    } else {
+                        debug!("write: No TLS Protocol negotiated")
+                    }
+                }
+                if self.client_session.wants_read() {
+                    let count = self.client_session.read_tls(&mut self.tcpstream)?;
+                    trace!("std::io::Write::write(): read_tls() -> {}", count);
+                    if count == 0 {
+                        break;
+                    }
+                    if let Err(err) = self.client_session.process_new_packets() {
+                        return Err(std::io::Error::new(std::io::ErrorKind::Other, err));
+                    }
+                } else {
+                    break;
+                }
+            }
         }
 
         trace!("std::io::Write::write() done");
@@ -107,32 +111,30 @@ impl std::io::Write for TlsTcpStream {
 
     fn flush(&mut self) -> std::io::Result<()> {
         trace!("std::io::Write::flush()");
-        let mut tlssession = self.tlssession.lock().unwrap();
-
         loop {
-            while tlssession.wants_write() {
-                let count = tlssession.write_tls(&mut self.tcpstream)?;
+            while self.client_session.wants_write() {
+                let count = self.client_session.write_tls(&mut self.tcpstream)?;
                 trace!("std::io::Write::flush(): wrote tls ({})", count);
             }
 
-            if self.is_handshaking && !tlssession.is_handshaking() {
+            if self.is_handshaking && !self.client_session.is_handshaking() {
                 self.is_handshaking = false;
 
-                if let Some(protocol) = tlssession.get_protocol_version() {
-                    debug!("Protocol {:?} negotiated", protocol)
+                if let Some(protocol) = self.client_session.get_protocol_version() {
+                    debug!("flush: Protocol {:?} negotiated", protocol)
                 } else {
-                    debug!("No TLS Protocol negotiated")
+                    debug!("flush: No TLS Protocol negotiated")
                 }
             }
 
-            if tlssession.wants_read() {
-                let count = tlssession.read_tls(&mut self.tcpstream)?;
+            if self.client_session.wants_read() {
+                let count = self.client_session.read_tls(&mut self.tcpstream)?;
                 trace!("std::io::Write::flush(): read_tls() -> {}", count);
                 if count == 0 {
                     break;
                 }
 
-                if let Err(err) = tlssession.process_new_packets() {
+                if let Err(err) = self.client_session.process_new_packets() {
                     return Err(std::io::Error::new(std::io::ErrorKind::Other, err));
                 }
             } else {
@@ -145,20 +147,18 @@ impl std::io::Write for TlsTcpStream {
     }
 }
 
-impl std::io::Read for TlsTcpStream {
+impl std::io::Read for Stream {
     fn read(&mut self, buffer: &mut [u8]) -> std::io::Result<usize> {
         trace!("std::io::Read::read() with buf size {}", buffer.len());
-        let mut tlssession = self.tlssession.lock().unwrap();
-
         loop {
-            if tlssession.wants_read() {
-                let count = tlssession.read_tls(&mut self.tcpstream)?;
+            if self.client_session.wants_read() {
+                let count = self.client_session.read_tls(&mut self.tcpstream)?;
                 trace!("transfer_read(): read_tls() -> {}", count);
                 if count == 0 {
                     break;
                 }
 
-                if let Err(err) = tlssession.process_new_packets() {
+                if let Err(err) = self.client_session.process_new_packets() {
                     return Err(std::io::Error::new(std::io::ErrorKind::Other, err));
                 }
             } else {
@@ -166,7 +166,7 @@ impl std::io::Read for TlsTcpStream {
             }
         }
 
-        let read_bytes = tlssession.read(&mut buffer[..])?;
+        let read_bytes = self.client_session.read(&mut buffer[..])?;
         trace!("std::io::Read::read() done");
         Ok(read_bytes)
     }