+++ src/lib.rs
@@ -33,6 +33,8 @@
 #![allow(clippy::must_use_candidate)]
 
 #[macro_use]
+extern crate lazy_static;
+#[macro_use]
 extern crate log;
 #[macro_use]
 extern crate serde_derive;